LPD-30592 Extract language.json tracker to its own class to make things more readable
Since we need to add two more trackers in the following commits.
